refactor(auth): Centraliza useAuth hook y mejora respuestas API
- Mueve la definición del hook useAuth de AuthContext.tsx a su propio archivo src/hooks/useAuth.ts.
- Exporta useAuth desde src/hooks/index.ts para permitir importaciones centralizadas
- Actualiza las importaciones de useAuth en Navbar.tsx para usar la nueva ruta.
- Introduce funciones helper apiErrorResponse y apiSuccessResponse en src/lib/utils.ts para estandarizar las respuestas JSON en las API Routes.
- Implementa el uso de apiErrorResponse y apiSuccessResponse en /api/auth/session y /api/auth/logout.
