+++ src/components/layout/Navbar.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Link from 'next/link';
-import { useAuth } from '@/contexts/AuthContext'; 
+import { useAuth } from '@/hooks'
 import { Button } from '@/components/ui/button'; 
 import { Skeleton } from '@/components/ui/skeleton'
