+++ src/hooks/useAuth.ts
@@ -1,1 +1,10 @@
-export {};
+import { useContext } from 'react';
+import { AuthContext, AuthContextType } from '@/contexts/AuthContext'; 
+
+export function useAuth() {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error('useAuth debe ser usado dentro de un AuthProvider');
+  }
+  return context;
+}
